ordem-servico-corretiva/os-report/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.9
python-multipart==0.0.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
aiosqlite==0.18.0 